requirements.txt
flask==2.0.1
openai>=1.0.0
python-dotenv==0.19.0
neo4j==4.4.3
redis==4.5.1
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
plotly==5.3.1
dash==2.0.0
networkx==2.6.3
python-jose==3.3.0
pytest==6.2.5
gunicorn==20.1.0
requests==2.31.0 